Changed verbose-file-{entry,exit}() arguments
verbose-file-{entry,exit}() functions now take an additional argument at
the beginning that is the filename/path of the file the function is
being called for. This used to be taken from $BASH_SOURCE but that
reported the name of the function file instead of the file the function
was being called from.